Adds a custom git commit message
Adds a custom git commit template that explains how to write a
commit message that is consise, consistent, and communicates context
about changes. See also http://chris.beams.io/posts/git-commit/ and
the 7 rules of a great commit message which are...

Separate subject from body with a blank line
Limit the subject line to 50 characters
Capitalize the subject line
Do not end the subject line with a period
Use the imperative mood in the subject line
Wrap the body at 72 characters
Use the body to explain what and why vs. how